I have a problem with loading users from server
