+++ src/app/store.ts
@@ -1,10 +1,12 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
 // eslint-disable-next-line import/no-cycle
 import counterReducer from '../features/counter/counterSlice';
+import usersReducer from '../features/users/users';
 
 export const store = configureStore({
   reducer: {
     counter: counterReducer,
+    users: usersReducer,
   },
 });
