+++ src/components/users/users.tsx
@@ -1,0 +1,9 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+const usersSlice = createSlice({
+  name: 'users',
+  initialState: [],
+  reducers: {
+    
+  }
+})
